PR Into OSLabs

+++ backend/server.ts
@@ -1,75 +1,70 @@
-import express, { Request, Response } from 'express';
+import express, { NextFunction, Request, Response } from 'express';
 import fs from 'fs';
-import cors from 'cors';
-import { exec } from 'child_process';
-import * as path from 'path';
-import { ServerError, GlobalErrorObject } from './backend-types';
-const cookieParser = require('cookie-parser');
+import { ServerError } from './backend-types';
+import process from 'process';
+import cookieParser from 'cookie-parser';
 
 // const PORT = process.env.PORT || 3003;
-const SOCKETFILE = '/run/guest-services/backend.sock';
+let SOCKETFILE: string;
+if (process.env.MODE === 'browser') {
+  SOCKETFILE = '3000';
+} else {
+  SOCKETFILE = '/run/guest-services/backend.sock';
+
+  // Resets the docker socket, prevents the issue of VM socket being in use error but not always
+  try {
+    fs.unlinkSync(SOCKETFILE);
+    console.log('Deleted the UNIX file.');
+  }
+  catch (err) {
+    console.log('Did not need to delete the UNIX socket file.');
+  }
+}
+
+
+
 const app = express();
 
-// Resets the docker socket, prevents the issue of VM socket being in use error but not always
-try {
-  fs.unlinkSync(SOCKETFILE);
-  console.log('Deleted the UNIX file.');
-}
-catch (err) {
-  console.log('Did not need to delete the UNIX socket file.');
-}
-
-app.use(cors());
 app.use(cookieParser());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-// import apiRouter from './server/routes/apiRouter';
-import commandRouter from './server/routes/commandRouter';
-// import initRouter from './server/routes/initRouter';
-// import loginRouter from './server/routes/loginRouter';
-// import logoutRouter from './server/routes/logoutRouter';
-// import setupRouter from './server/routes/setupRouter';
-// import signupRouter from './server/routes/signupRouter';
+import containerRouter from './routers/docker/containerRouter';
+import imageRouter from './routers/docker/imageRouter';
+import volumeRouter from './routers/docker/volumeRouter';
+import networkRouter from './routers/docker/networkRouter';
+import systemRouter from './routers/docker/systemRouter';
+import configRouter from './routers/prometheus/configRouter';
 
-app.use(express.static('SetupApp'));
-
-// Defining routers...
-// app.use('/k8', (req: Request, res: Response) => {
-//   res.status(200).sendFile(path.join(__dirname, '../SetupApp/index.html'));
-// });
-
-
-// app.use('/gapi', apiRouter);
-app.use('/command', commandRouter);
-// app.use('/init', initRouter);
-// app.use('/login', loginRouter);
-// app.use('/logout', logoutRouter);
-// app.use('/setup', setupRouter);
-// app.use('/signup', signupRouter);
+app.use('/api/docker/container', containerRouter);
+app.use('/api/docker/image', imageRouter);
+app.use('/api/docker/volume', volumeRouter);
+app.use('/api/docker/network', networkRouter);
+app.use('/api/docker/system', systemRouter);
+app.use('/api/prometheus/config', configRouter);
 
 // Handling requests to unknown endpoints...
 app.use('/', (req: Request, res: Response): Response => {
   return res
     .status(404)
-    .send({ error: 'Unknown endpoint YES HIT ROUTE please try again.' });
+    .json({ error: 'Endpoint does not exist' });
 });
 
 // Handling global errors...
-app.get(
-  '/',
+app.use(
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  (err: ServerError, req: Request, res: Response): Response => {
-    const defaultErr: GlobalErrorObject = {
+  (err: ServerError, req: Request, res: Response, next: NextFunction): Response => {
+    const defaultErr: ServerError = {
       log: 'Express error handler caught unknown middleware error',
       status: 500,
       message: { err: 'An error occurred' },
     };
-    const errorObj: ServerError = Object.assign(defaultErr, err);
+    const errorObj: ServerError = Object.assign({}, defaultErr, err);
+    console.log(errorObj.log);
     return res.status(errorObj.status).json(errorObj.message);
   }
 );
 
 app.listen(SOCKETFILE, (): void => {
-  console.log(`Listening on port ${SOCKETFILE}`);
+  console.log(`Listening on socket: ${SOCKETFILE}`);
 });
